added role middleware to the controllers, functions and routes

+++ app/Http/Controllers/Admins/RoleController.php
@@ -10,6 +10,10 @@ use Inertia\Inertia;
 
 class RoleController extends Controller
 {
+    public function __construct() {
+        $this->middleware(['role:super-admin|admin|moderator']);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -39,16 +43,19 @@ class RoleController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request) {
-        $this->validate($request, [
-            'name' => ['required', 'max:25', 'unique:roles'],
-            'permissions' => 'required'
-        ]);
-        $role = Role::create([
-            'name' => $request->name,
-            'guard_name' => 'web',
-        ]);
-        if ($request->has('permissions')) {
-            $role->givePermissionTo(collect($request->permissions)->pluck('id')->toArray());
+        if (auth()->user()->hasAnyRole(['super-admin', 'admin'])) {
+            $this->validate($request, [
+                'name' => ['required', 'max:25', 'unique:roles'],
+                'permissions' => 'required'
+            ]);
+            $role = Role::create([
+                'name' => $request->name,
+                'guard_name' => 'web',
+            ]);
+            if ($request->has('permissions')) {
+                $role->givePermissionTo(collect($request->permissions)->pluck('id')->toArray());
+            }
+            return back();
         }
         return back();
     }
@@ -83,15 +90,18 @@ class RoleController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Role $role) {
-        $this->validate($request, [
-            'name' => ['required', 'max:25'],
-            'permissions' => 'required'
-        ]);
-        if ($request->has('permissions')) {
-            $role->givePermissionTo(collect($request->permissions)->pluck('id')->toArray());
+        if (auth()->user()->hasAnyRole(['super-admin', 'admin'])) {
+            $this->validate($request, [
+                'name' => ['required', 'max:25'],
+                'permissions' => 'required'
+            ]);
+            if ($request->has('permissions')) {
+                $role->givePermissionTo(collect($request->permissions)->pluck('id')->toArray());
+            }
+            $role->syncPermissions(collect($request->permissions)->pluck('id')->toArray());
+            $role->update(['name' => $request->name]);
+            return back();
         }
-        $role->syncPermissions(collect($request->permissions)->pluck('id')->toArray());
-        $role->update(['name' => $request->name]);
         return back();
     }
 
@@ -102,7 +112,10 @@ class RoleController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Role $role) {
-        $role->delete();
+        if (auth()->user()->hasAnyRole(['super-admin', 'admin'])) {
+            $role->delete();
+            return back();
+        }
         return back();
     }
 }
